sub admin final

+++ app/Http/Controllers/Admin/AppliedCareerController.php
@@ -1,7 +1,7 @@
 <?php
 
 namespace App\Http\Controllers\Admin;
-
+use Illuminate\Support\Facades\Gate;
 use App\Http\Controllers\Controller;
 use App\Models\AppliedCareer;
 use Illuminate\Http\Request;
@@ -13,6 +13,7 @@ class AppliedCareerController extends Controller
      */
     public function index()
     {
+        abort_unless(Gate::allows('applied-career'), 403);
         $title = 'Careers';
         $sub_title = 'Careers';
 
@@ -21,12 +22,14 @@ class AppliedCareerController extends Controller
 
     public function show(Request $request)
     {
+        abort_unless(Gate::allows('applied-career'), 403);
         $data = AppliedCareer::getFullData($request);
         return $data;
     }
 
     public function destroy(Request $request)
     {
+        abort_unless(Gate::allows('applied-career-delete'), 403);
         $delete = AppliedCareer::deleteData($request);
         return response()->json($delete);
     }
