+++ app/Http/Controllers/Admin/CareerBranchController.php
@@ -1,7 +1,7 @@
 <?php
 
 namespace App\Http\Controllers\Admin;
-
+use Illuminate\Support\Facades\Gate;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\StoreCareerBranchRequest;
 use App\Http\Requests\Admin\UpdateCareerBranchRequest;
@@ -13,6 +13,7 @@ class CareerBranchController extends Controller
 {
     public function index()
     {
+        abort_unless(Gate::allows(' career-branches'), 403);
         $title = 'Career Branch';
         $sub_title = 'Career Branch';
 
@@ -21,6 +22,7 @@ class CareerBranchController extends Controller
 
     public function create()
     {
+        abort_unless(Gate::allows(' career-branches-create'), 403);
         $title = 'Career Branch';
         $sub_title = 'Add';
         return view('admin.career-branch.create',compact('title','sub_title'));
@@ -28,6 +30,7 @@ class CareerBranchController extends Controller
 
     public function store(StoreCareerBranchRequest $request)
     {
+        abort_unless(Gate::allows(' career-branches-create'), 403);
         try{
             $save= CareerBranch::createData($request);
 
@@ -56,12 +59,14 @@ class CareerBranchController extends Controller
 
     public function show(Request $request)
     {
+        abort_unless(Gate::allows(' career-branches'), 403);
         $data = CareerBranch::getFullData($request);
         return $data;
     }
 
     public function edit($id)
     {
+        abort_unless(Gate::allows(' career-branches-edit'), 403);
         $data = CareerBranch::getData($id);
         if(!$data){
             abort(404);
@@ -73,6 +78,7 @@ class CareerBranchController extends Controller
 
     public function update(UpdateCareerBranchRequest $request, $id)
     {
+        abort_unless(Gate::allows(' career-branches-edit'), 403);
         try{
             $save= CareerBranch::updateData($request);
 
@@ -101,6 +107,7 @@ class CareerBranchController extends Controller
 
     public function destroy(Request $request)
     {
+        abort_unless(Gate::allows(' career-branches-delete'), 403);
         $delete = CareerBranch::deleteData($request);
         return response()->json($delete);
     }
