+++ app/Http/Controllers/Admin/CareerDepartmentController.php
@@ -1,7 +1,7 @@
 <?php
 
 namespace App\Http\Controllers\Admin;
-
+use Illuminate\Support\Facades\Gate;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\StoreCareerDepartmentRequest;
 use App\Http\Requests\Admin\UpdateCareerDepartmentRequest;
@@ -13,6 +13,7 @@ class CareerDepartmentController extends Controller
 {
     public function index()
     {
+        abort_unless(Gate::allows('careers-departments'), 403);
         $title = 'Career Department';
         $sub_title = 'Career Department';
 
@@ -21,6 +22,7 @@ class CareerDepartmentController extends Controller
 
     public function create()
     {
+        abort_unless(Gate::allows('careers-departments-create'), 403);
         $title = 'Career Branch';
         $sub_title = 'Add';
         return view('admin.career-department.create',compact('title','sub_title'));
@@ -28,6 +30,7 @@ class CareerDepartmentController extends Controller
 
     public function store(StoreCareerDepartmentRequest $request)
     {
+        abort_unless(Gate::allows('careers-departments-create'), 403);
         try{
             $save= CareerDepartment::createData($request);
 
@@ -56,12 +59,14 @@ class CareerDepartmentController extends Controller
 
     public function show(Request $request)
     {
+        abort_unless(Gate::allows('careers-departments'), 403);
         $data = CareerDepartment::getFullData($request);
         return $data;
     }
 
     public function edit($id)
     {
+        abort_unless(Gate::allows('careers-departments-edit'), 403);
         $data = CareerDepartment::getData($id);
         if(!$data){
             abort(404);
@@ -73,6 +78,7 @@ class CareerDepartmentController extends Controller
 
     public function update(UpdateCareerDepartmentRequest $request, $id)
     {
+        abort_unless(Gate::allows('careers-departments-edit'), 403);
         try{
             $save= CareerDepartment::updateData($request);
 
@@ -101,6 +107,7 @@ class CareerDepartmentController extends Controller
 
     public function destroy(Request $request)
     {
+        abort_unless(Gate::allows('careers-departments-delete'), 403);
         $delete = CareerDepartment::deleteData($request);
         return response()->json($delete);
     }
