+++ app/Http/Controllers/Admin/EligibilityCheckController.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace App\Http\Controllers\Admin;
+use Illuminate\Support\Facades\Gate;
 
 use App\Http\Controllers\Controller;
 use App\Models\EligibilityCheck;
@@ -10,6 +11,7 @@ class EligibilityCheckController extends Controller
 {
     public function index()
     {
+        abort_unless(Gate::allows('eligibility-check'), 403);
         $title = 'Eligibility Check';
         $sub_title = 'Eligibility Check';
 
@@ -18,12 +20,14 @@ class EligibilityCheckController extends Controller
 
     public function show(Request $request)
     {
+        abort_unless(Gate::allows('eligibility-check'), 403);
         $data = EligibilityCheck::getData($request);
         return $data;
     }
 
     public function edit($id)
     {
+        abort_unless(Gate::allows('eligibility-check-edit'), 403);
         $data = EligibilityCheck::getFullData($id);
 
         return response()->json(['data'=>$data,'status'=>true]);
@@ -31,6 +35,7 @@ class EligibilityCheckController extends Controller
 
     public function destroy(Request $request)
     {
+        abort_unless(Gate::allows('eligibility-check-delete'), 403);
         $delete = EligibilityCheck::deleteData($request);
         return response()->json($delete);
     }
