+++ app/Http/Controllers/Admin/ExploreController.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace App\Http\Controllers\Admin;
+use Illuminate\Support\Facades\Gate;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\StoreExploreRequest;
@@ -13,6 +14,7 @@ class ExploreController extends Controller
 {
     public function index()
     {
+        abort_unless(Gate::allows('explore'), 403);
         $title = 'Explore';
         $sub_title = 'Explore';
 
@@ -20,7 +22,7 @@ class ExploreController extends Controller
     }
 
     public function create()
-    {
+    {  abort_unless(Gate::allows('explore-create'), 403);
         $title = 'Explore';
         $sub_title = 'Add';
         return view('admin.explore.create',compact('title','sub_title'));
@@ -28,6 +30,7 @@ class ExploreController extends Controller
 
     public function store(StoreExploreRequest $request)
     {
+         abort_unless(Gate::allows('explore-create'), 403);
         try{
             $save= Explore::createData($request);
 
@@ -56,12 +59,14 @@ class ExploreController extends Controller
 
     public function show(Request $request)
     {
+         abort_unless(Gate::allows('explore'), 403);
         $data = Explore::getFullData($request);
         return $data;
     }
 
     public function edit($id)
     {
+        abort_unless(Gate::allows('explore-edit'), 403);
         $data = Explore::getData($id);
         if(!$data){
             abort(404);
@@ -73,6 +78,7 @@ class ExploreController extends Controller
 
     public function update(UpdateExploreRequest $request, $id)
     {
+        abort_unless(Gate::allows('explore-edit'), 403);
         try{
             $save= Explore::updateData($request);
 
@@ -101,6 +107,7 @@ class ExploreController extends Controller
 
     public function destroy(Request $request)
     {
+        abort_unless(Gate::allows('explore-delete'), 403);
         $delete = Explore::deleteData($request);
         return response()->json($delete);
     }
